+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `BackgroundWithID` creates a new context inheriting the request ID.
 
 ## [1.3.0] - 2016-09-02
 ### Added